requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
psycopg[binary]==3.2.9
